+++ supreme/common/src/jvmTest/kotlin/DebugStatementParserTest.kt
@@ -9,6 +9,7 @@ import at.asitplus.testballoon.withData
 import com.android.keyattestation.verifier.KeyDescription
 import com.google.android.attestation.ParsedAttestationRecord
 import de.infix.testBalloon.framework.core.testSuite
+import io.kotest.matchers.nulls.shouldBeNull
 import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.shouldBe
 import kotlinx.serialization.json.Json
@@ -54,7 +55,7 @@ val DebugStatementParserTest by testSuite {
         return@testSuite
     }
 
-    withData(proofs) - {
+    withData(proofs.withIndex()) - { (index, it) ->
         val attestationCertChain =
             it.map {
                 val certBytes = Base64.getUrlDecoder().decode(it)
@@ -71,6 +72,9 @@ val DebugStatementParserTest by testSuite {
 
 
             val androidAttestationExtension = attestationCertChain.first().androidAttestationExtension
+            "from chain should be same as from leaf" {
+                androidAttestationExtension shouldBe attestationCertChain.androidAttestationExtension
+            }
             "convert" - {
                 withData(nameFn = { it.subjectX500Principal.toString() }, attestationCertChain) {
                     it.toKmpCertificate().isSuccess shouldBe true
@@ -83,6 +87,28 @@ val DebugStatementParserTest by testSuite {
                     own.encodeToDer() shouldBe it.encoded
                 }
             }
+
+            if (index > 0) {
+                val prev = proofs[index - 1].map {
+                    val certBytes = Base64.getUrlDecoder().decode(it)
+                    catchingUnwrapped {
+                        certificateFactory.generateCertificate(
+                            ByteArrayInputStream(
+                                certBytes
+                            )
+                        ) as X509Certificate
+                    }.getOrNull()
+                }.filterNotNull()
+
+                if (prev.isNotEmpty() && ((androidAttestationExtension != null) && prev.androidAttestationExtension != null)) {
+                    "concatenating two chains (current = $index) should get the same exnt as just from the current chain" {
+                        //leaf = fist in chain, root = last in chain. so we went the closest to CURRENT root. hence we need to prepent
+                        //s.t. the chain gets extended below the leaf and not above the root
+                        (prev + attestationCertChain).androidAttestationExtension shouldBe attestationCertChain.androidAttestationExtension
+                    }
+                }
+            }
+
             val fromGoogle =
                 catchingUnwrapped { ParsedAttestationRecord.createParsedAttestationRecord(attestationCertChain) }.getOrNull()
 
@@ -90,6 +116,8 @@ val DebugStatementParserTest by testSuite {
                 if (fromGoogle == null) {
                     System.err.println("Old Google parser glitched out")
                     val newParser = catchingUnwrapped { KeyDescription.parseFrom(attestationCertChain.first()) }
+
+                    attestationCertChain.androidAttestationExtension.shouldBeNull()
                     return@invoke//well, well, well…}
                 }
                 androidAttestationExtension.shouldNotBeNull()